+++ changelog.md
@@ -3,6 +3,11 @@ title: changelog
 layout: page
 ---
 
+# v0.1.2b
+*19/12/2021*
+- Consistency between loading one post and multiple
+- Show author and date on post preview
+
 # v0.1.1b
 *19/12/2021*
 - Redirect to board after posting